% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspt.R
\name{aspt}
\alias{aspt}
\title{aspt}
\usage{
aspt(x)
}
\arguments{
\item{x}{data.frame as specified in details.}
}
\description{
Questa funzione permette il calcolo dell'Average Score Per Taxon (ASPT).
}
\details{
Il calcolo viene effettuato utilizzando i punteggi definiti da Davy-Bowker et al. 2008. È possibile visionare i punteggi digitando aspt_v e premendo invio.
}
\examples{
data(esempio_su)
surber <- data.prep(esempio_su)
aspt(surber)
}
\references{
Davy-Bowker J., Clarke R., Corbin T., Vincent H, Pretty J., Hawczak A., Blackburn J., Murphy J., Jones I., 2008. River Invertebrate Classification Tool. Final report. WFD72C. SNIFFER. 276 pp
}
\keyword{ASPT}

