% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ept.R
\name{ept}
\alias{ept}
\title{ept}
\usage{
ept(x)
}
\arguments{
\item{x}{data.frame as specified in details.}
}
\description{
Questa funzione permette il calcolo dell'EPT (n° di famiglie di Efemerotteri, Plecotteri e Tricotteri)
}
\examples{
data(esempio_su)
surber <- data.prep(esempio.su)
ept(surber)
}
\keyword{EPT}

