% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eptd.R
\name{EPTD}
\alias{EPTD}
\title{EPTD}
\usage{
EPTD(x)
}
\arguments{
\item{un}{oggetto di classe STAR o ISA}
}
\description{
Questa funzione permette il calcolo del logartimo degli EPTD selezionati (Efemerotteri, Plecotteri, Tricotteri ed Efemerotteri)
}
\details{
La metrica EPTD viene calcolata utilizzando la formula
log10(Sel_EPTD+1)
dove log10 rappresenta il logaritmo in base 10 e SEL_EPTD l'abbondanza degli EPTD selezionati.
}
\examples{
load(esempio_su)
surber <- data.prep(esempio_sub)
EPTD(surber)
# lista degli EPTD
rownames(aspt_v[aspt_v$EPTD==1,])
}
\keyword{EPTD}
