% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOLD.R
\name{gold}
\alias{gold}
\title{1-GOLD}
\usage{
gold(x)
}
\arguments{
\item{x}{a data.prep object.}
}
\description{
This function calculates the metric 1-GOLD.
}
\examples{
data(oglio)
oglio.prep <- data.prep(oglio)
gold(oglio.prep)
}
\keyword{dati}
\keyword{dei}
\keyword{preparazione}
