% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limeco.R
\name{LIMeco}
\alias{LIMeco}
\title{LIMeco}
\usage{
LIMeco(X, STA = NULL, O2 = NULL, NH4 = NULL, NO3 = NULL, TP = NULL,
  writexlsx = F)
}
\arguments{
\item{X}{data.frame as specified in details.}

\item{STA}{column number of site names.}

\item{O2}{column number of oxygen data.}

\item{NH4}{column number of ammonium data.}

\item{NO3}{column number of nitrate data.}

\item{TP}{column number of total phosporus data.}
}
\description{
Calculate LIMeco.
}
\examples{
year <- c(rep(2013,5), rep(2014,5))
STA <- rep(letters[1:2],5)
O2 <- rpois(10, 100)
NO3 <- abs(rnorm(10, 1.3, 0.5))
NH4 <- abs(rnorm(10, 0.5, 0.2))
TP <- abs(rnorm(10, 2.5, 1))
ex <- data.frame(year,STA,O2,NO3,NH4,TP)
ex
data <- LIMeco(ex,2,3,4,5,6)
# save in xlsx format, data will be stored in working directory (not run)
# write.xlsx(data, "limeco.xlsx")
tapply(data$LIMeco, list(ex$STA, ex$year), mean)
}
\keyword{limECO}
