% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts}
\alias{mts}
\title{mts}
\usage{
mts(x)
}
\arguments{
\item{data.frame}{un oggetto di classe STAR o ISA}
}
\description{
Questa funzione permette il calcolo del Mayfly Total Score (Buffagni et al. 1997).
}
\details{
Se Siphlonurus risulta essere l'unico efemerottero nella comunità allora il suo punteggio passa da 3 a 1.
}
\examples{
macropen()
}
\references{
Buffagni A., 1997 Mayfly community composition and the biological quality of streams. In Landolt P. & Sartori M. (eds), Ephemeroptera & Plecoptera: Biology-Ecology-Systematics. MTL, Fribourg: 235-246
}
\keyword{MTS}
