% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfam.R
\name{nfam}
\alias{nfam}
\title{nfam}
\usage{
nfam(x)
}
\arguments{
\item{data.frame}{as specified in details.}
}
\description{
Calculate nfam.
}
\examples{
nfam()
}
\keyword{data}
\keyword{preparation}

