% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooao.R
\name{ooao}
\alias{ooao}
\title{ooao}
\usage{
ooao(LIMeco, STAR_ICMi, cex)
}
\arguments{
\item{data.frame}{as specified in details.}
}
\description{
Calculate ooao.
}
\examples{
ooao()
}
\keyword{data}
\keyword{preparation}
