% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{shannon}
\usage{
shannon(x)
}
\arguments{
\item{data.frame}{as specified in details.}
}
\description{
Calculate shannon.
}
\examples{
shannon()
}
\keyword{data}
\keyword{preparation}
