% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspt.R
\name{aspt}
\alias{aspt}
\title{aspt}
\usage{
aspt(x)
}
\arguments{
\item{x}{a data.prep object.}
}
\description{
This function calculates the Average Score Per Taxon (ASPT).
}
\details{
Calculation made according to Davy-Bowker et al. 2008. Scores can be viewed by typing aspt_v and pressing enter.
}
\examples{
data(oglio)
oglio.prep <- data.prep(oglio)
aspt(oglio.prep)
}
\references{
Davy-Bowker J., Clarke R., Corbin T., Vincent H, Pretty J., Hawczak A., Blackburn J., Murphy J., Jones I., 2008. River Invertebrate Classification Tool. Final report. WFD72C. SNIFFER. 276 pp
}
\seealso{
\code{\link{data.prep}}
}
\keyword{ASPT}
