% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ept.R
\name{ept}
\alias{ept}
\title{ept}
\usage{
ept(x)
}
\arguments{
\item{x}{a data.prep object.}
}
\description{
This function calculates the EPT metric (number of Plecoptera, Eèhemeroptera and Trichoptera families)
}
\examples{
data(oglio)
oglio.prep <- data.prep(oglio)
ept(oglio)
}
\keyword{EPT}
