% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOLD.R
\name{gold}
\alias{gold}
\title{1-GOLD}
\usage{
gold(x)
}
\arguments{
\item{x}{un oggetto di classe ISA O STAR.}
}
\description{
Calculate 1-GOLD.
}
\examples{
1-GOLD()
}
\keyword{dati}
\keyword{dei}
\keyword{preparazione}

